requirements.txt
PyYAML
pydantic
pydantic-settings
boto3